+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Added
 
 - `limestone/Chips.Chip` prop `multiline` to be able to show the text on multiple lines
+- `limestone/IconItem` prop `description`
+- `limestone/ImageItem` prop `secondaryLabel`
 - `limestone/Item` prop `secondaryLabel`
 - `limestone/Slider` props `ticks` and `labels` to display equally spaced tick marks and labels
 - `limestone/Slider` prop `alignStepsWithTicks` to snap the knob to tick marks
@@ -28,7 +30,6 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ### Added
 
-- `limestone/IconItem` prop `description`
 - `limestone/Image` props `backgroundColor` and `backgroundSrc` to be able to change the background of a transparent image
 - `limestone/ImageItem` props `backgroundColor` and `backgroundSrc` to be able to change the background of a transparent image