+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Fixed
 
 - `limestone/ContextualPopupDecorator` hole punch scrim to reposition together with the popup when the wrapped component is updated
+- `limestone/VirtualList` long press scroll between multiple VirtualLists
 
 ## [1.11.1] - 2026-09-09